+++ tests/test_fixlengths.rs
@@ -28,7 +28,7 @@ fn prop_fixlengths_all_maxlen() {
 
 #[test]
 fn prop_fixlengths_explicit_len() {
-    fn p(rows: Vec<CsvRecord>, expected_len: uint) -> TestResult {
+    fn p(rows: Vec<CsvRecord>, expected_len: usize) -> TestResult {
         if expected_len == 0 || rows.is_empty() {
             return TestResult::discard();
         }
@@ -44,5 +44,5 @@ fn prop_fixlengths_explicit_len() {
         for r in got.iter() { assert_eq!(r.as_slice().len(), got_len) }
         TestResult::from_bool(rassert_eq!(got_len, expected_len))
     }
-    qcheck(p as fn(Vec<CsvRecord>, uint) -> TestResult);
+    qcheck(p as fn(Vec<CsvRecord>, usize) -> TestResult);
 }
